Rename pool into arena
