+++ kernel-rs/src/arena.rs
