+++ kernel-rs/src/kernel.rs
@@ -6,6 +6,7 @@ use core::{
 use spin::Once;
 
 use crate::{
+    arena::RcArena,
     bio::Bcache,
     console::{consoleinit, Console},
     file::{Devsw, File},
@@ -15,7 +16,6 @@ use crate::{
     page::Page,
     param::{NCPU, NDEV, NFILE, NINODE},
     plic::{plicinit, plicinithart},
-    pool::RcArena,
     println,
     proc::{cpuid, procinit, scheduler, Cpu, ProcessSystem},
     riscv::PGSIZE,
